added first application run configuration dialog
if running the application for the first time the user must set an library
http server address and username and password for his account. For now the
logic at launch is that the application will not start if there if no network
access, and if there is route to library server url the application will
start a configuration dialog. When the correct http server address is entered
the splash screen will start and IDL and orgTree files will be downloaded
and parsed