Step 1 in IDL speedup: Use the fm_IDL.xml from assets instead of downloading; shrink fm_IDL.xml by removing unused elements.  Under the emulator parsing the IDL is ridiculously slow (35s).  Still loads of work to do; it's only slightly faster on the emulator (30s) and still noticeably slow on hardware (~5s).
Removed openils_idl.jar and instead put the modified sources under
Android/src/java/org/open_ils/.  Nat at least it's obvious that it's
modified version of the code in Open-ILS/src/java.  It's modified because
the javax.xml.stream parser is not available on Android and was replaced
with an xmlpull parser.