Improvements surrounding organizations.
Make initialization faster by (1) parsing OrgTree.js organization array as
JSON.

Do not include in the library list organizational units that are marked as
not visible in OPAC.  This reduces the list from 338 items to 181 items.
Improved readability of the list by removing space-padding that is done
for the web drop-down list, and putting the spinner in dialog mode.

Sort organizations by name, with (1) home library selected, and (2) root
organization first.  This narrows the results of default searches.