Added an android authenticator based on Apache 2.0 licensed code from https://github.com/Udinic/AccountAuthenticator
The main problem in the EG app in the android-minor-cleanup branch is that when the auth token expires or you rotate the device at the wrong time you get repeated crashes.  My hypothesis is that using native android.accounts facilities for getting access to auth tokens will clean up the auth handling and also make authentication accessible to Services.