Not yet successful attempt to move biblio.record.copy_count out of main search.
That call is taking 30% of the total search time, and it is not even necessary until you look at the item details.  Made sure that the orgID is properly passed from SearchCatalog through to BasicDetailsFragment.  Not working yet because my first attempt was to call getCopyCount on the main thread.  End goal: volley-based OSRF response handler.