Implemented onClick handling with RecyclerView
I think the activity handling, e.g. launching the details activity
(SampleUnderlinesNoFade) should be handled in the activity.  That frees
the search results fragment to be free to be reused in e.g. a future
tablet search activity.  So the fragment now implements an
OnRecordClickListener that is fed from the RecordViewAdapter.