Pass back the resultsList from details activity to search activity.
This interface is a bit strange, but here's the deal.  Both the search
activity and the details activity can "search" and modify the set of
records.  The details activity can pass back the augmented results in the
activity result Intent.  It used to pass that back only when you hit the
"back" button but not when you hit the "home" or "up" button in the action
bar.  Not it does it for both.
# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# Explicit paths specified without -i or -o; assuming --only paths...
# On branch try-recyclerview
# Changes to be committed:
#	modified:   core/src/org/evergreen_ils/searchCatalog/SearchCatalogListView.java
#
# Changes not staged for commit:
#	modified:   core/src/org/evergreen_ils/searchCatalog/SampleUnderlinesNoFade.java
#	modified:   core/src/org/evergreen_ils/searchCatalog/SearchCatalog.java
#
# Untracked files:
#	constants.js
#	foo
#