Fixed NPE seen in play store
java.lang.NullPointerException: Attempt to invoke virtual method 'void org.evergreen_ils.searchCatalog.RecordInfo.setSearchFormat(java.lang.String)' on a null object reference
	at org.evergreen_ils.accountAccess.AccountAccess.getHolds(AccountAccess.java:620)
My conclusion is that this is a Librarian using the app inspecting
a hold that is not a T (title) or M (metarecord) hold.  I don't know how
to test one of those but I think this will fix the NPE.

Also removed some unused code that moved to the authenticator long ago.