Stashing work in progress on Navigation Drawer
It does (promise to) provide a quick way to access any point in the UI
without using back multiple times, but it is not a well thought out
strategy.  It duplicates the main menu in the drawer.  That may be useful
enough.  Then we will want the account switcher there too...