Fixed crashes caused by lifecycle management errors
When I moved the common if-not-init-then-restart code up into
BaseActivity, it meant that the superclass returned and the subclass
continued initializing, including referencing global state that was null.
So now subclasses of BaseActivity check to see if they are restarting and
skip the rest of onCreate if so.  A better long-term fix is to move the
sort of code that crashed into a ViewModel that is lifecycle aware.