Fix #15 - crash on failed renewal
renewCirc was crashing on an event (MAX_RENEWALS_REACHED) or list of
events (PATRON_EXCEEDS_FINES + PATRON_EXCEEDS_LOST_COUNT).  The functions
Utils.doRequest and GatewayRequest.recv() need to be refactored, but in
order to not change everything at once I've started using a
GatewayResponse object in select code paths.