Improve handling of online items
Acorn online items need marcxml data to determine the online locations.
But marcxml is bigger and takes longer to load than other metadata.  So
wait until the Item Details screen to load it, and disable the Place Hold
button until we do load it.  Also set a longer timeout for the marcxml
request.  Running the simulator with EDGE network reproduced the problem.

- don't load marcxml until Item Details is shown
- don't enable Place Hold or Online Access button until
  we get some clarity about what type of item it is
- increase network timeout for bre (marcxml) request