Add a random test of record IDs
I originally tried to use junit-quickcheck but didn't get it to work in
Android Studio.  I wanted a way to gather some random samples of records
and verify that my assumptions were true, e.g. about item_form.  I ended
up with this but I'll probably disable it, it's slow and does not
illustrate much.