Implement Located URI matching correctly
PINES shows only links to electronic resources that are owned by the
search org or an ancestor.  That is now implemented by
isVisibleViaLocatedURI().

Refactored other *AppBehavior classes to share as much code as practical,
and fixed NOBLEAppBehavior to treat online icon formats as online even for
cases where item_form is empty.