Clear activity stack when navigating home via nav drawer
Add flag Intent.FLAG_ACTIVITY_CLEAR_TOP when navigating to MainActivity.
This makes it work the same way as the iOS app: when hitting the
logo/home, all activities are cleared back to the main activity.  This
together with the prior commit (launchMode="singleTask") means that there
is only ever one MainActivity, at the top of the back stack.