Factor out GatewayJob.fetchAllOrgSettingsAsync
I'm not 100% happy with the naming or the fact that it uses GlobalScope,
but it factors out the fetching and loading of org settings in a similar
way to iOS with PromiseKit.  FinesActivity needs only settings for the
home org, which for PINES is significantly faster.

Move Organization into data now that it is Kotlin.