Exception handling WIP - FinesActivity
Figuring out how best to handle exceptions in coroutines.  Currently
launch propagates exceptions and the app crashes.  Async is supposed to
expose exceptions to the caller of await() so that may be worth trying.
Also there is supervisorScope.  The approach of the android-architecture
project was to create a sealed Result class and return that from every
suspend fun.  That seems to make the most sense to me.