Handle session expired errors with app restart on request
This is how iOS deals with session expired errors, and it's decent.  On
the plus side, the app is restarted cleanly and we have a new session
without jumping through any retry hoops.  On teh down side, you lose your
place in the app, which is especially annoying if you just finished a
search and you are browsing through 200 results.