Fix error handling of failed hold
The error when failing to create a hold is odd, it's a last_event obj
inside a result obj inside the payload obj.  Change the parseEvent func
to handle that too, that keeps the various error handling mechanisms all
within GatewayResult.create().  Converted Event to kotlin because some
things like this are just easier.