Extend RequestOptions to include cacheMaxTtlSeconds
And use it to limit the cache TTL of ORG_TREE_RETRIEVE to 24h.
Without this, org details that are part of that response (e.g. email
address) are cached for a very long time (1 year).

Create a new GatewayStringRequest to parallel GatewayJsonRequest
and implement max cache TTL in both.

Don't cache USER_FLESHED_RETRIEVE.