Switch to 4-digit versionName to match iOS
Except for CW MARS, which is at version 5.3.0, all other apps are
at 2.3.0.x and this will align with iOS version numbers.  This
will make it easier to write common Release Notes that give a
good summary of feature differences between production and latest,
for apps that do not release often.