Enhance LiveGatewayTest by logging into a session
Many APIs are not testable unless you login and call open-ils.auth.session.retrieve.
Now tests in this file can do that, and once you login and open a session, it is
automatically deleted at the end of the test run so the sessions do not consume
memory on the server.

Add a test using MULTICLASS_QUERY to flesh a bookbag.  It turns out that is how
the OPAC does it, and that way hides deleted records.