Do not clear search results on rotation
For this we store the accountId into the outState bundle;
if we have this and it matches the current account ID then
we just rotated and it is OK to keep the search results.
This check fixes the issue mentioned in commit 8f674d8b
that motivated clearing the results.