Use GlobalScope to avoid getting cancelled
Use GlobalScope to launch getAccount, because the lifecycleScope
is cancelled when LaunchActivity starts AuthenticatorActivity.
This fixes a bug where a new account required entering your
username + password twice.  But it still has a problem of calling
fetchSession twice, because LaunchActivity is recreated as a
result of calling AppCompatDelegate.setDefaultNightMode:
* AppCompatDelegate.setDefaultNightMode(YES)
* getAuthTokenFuture().await()
* ActivityCompat.recreate()
* onDestroy()
* onCreate()
* getAuthTokenFuture().await()
  --> Bundle
  --> Bundle
* fetchSession()
* fetchSession()

I wonder...was it ever thus?