Move versionName back to manifest
If you want the versionName in the build.gradle then you have to
have it in the app-specific build.gradle, cwmars_app/build.gradle.
At that point it is *less* convenient than having it in the manifest,
because you have to change 2 places and also the Play Console does
not correctly identify the versionName.