Tap on history item to see details
Because the history list can be quite large, and because
the items in the RecyclerView are async loaded, make the
tap show Item Details on only the one item.  The alternative
would have required (a) calling fetchCopyMODS for every
item in the list, and (b) risking a IllegalStateException
crash for passing too big a  list to RecordDetailsActivity.