Fix buttons on Place Hold and Hold Details
The button text was ellipsized at a fixed width on my Pixel 6.
Move the button row inside the ScrollView because as far as
I can tell there is no benefit to having it outside (at least
not as long as the whole thing is under a LinearLayout as
it is here...maybe a ConstraintLayout would work better).