Change hold status "Available" -> "Ready for pickup"
"Available" is the default in the OPAC, but that can be
confules with Copy Status of "Available" which means that
it is on the shelf.  PINES has customized the OPAC to
say "Ready for pickup" and that seems like the more sane
default.  If we need to change it back for some consortium