Handle auto-generated events like an alert block
The alert block causes a hold to fail with an event that was
not previously handled: "ilsevent":"","textcode":"STAFF_CHR".

Treat this as an event, and do not treat as an event the SUCCESS
"event" response from authentication.  Eliminate any handling of
the ilsevent field as an Int, since it can be empty, and we don't
really care about it.