Require accountType to change when using test server
The authenticator.xml declares an accountType, so you can't
derive it dynamically.  We want a different accountType for
the acorn test server, because if we use the same one, acorn
will try to reuse the saved auth and fail.  This requires the
user to delete the account from the device, which is inconvenient.