Fix bug causing a permissions dialog on every launch
The logic in the developer guide is not right, because
at least on API 34, ActivityCompat.shouldShowRequestPermissionRationale()
kept returning true every time.  So now we keep a denyCount and avoid
asking if it's greater than zero.  This works even if you disable
notifications through Settings, or later reenable them in Settings.

Also, remove the code in MainBaseActivity to handle launch intents;
they don't show up there.