Use constraint layout for grid button row
Originally I did this because I wanted to constrain the
ratio to 16:9 or 1.5, but I did not like that effect.
But it seems to work well with any of the buttons enabled,
so I kept it.  Also replaced the all-or-nothing debug
display of buttons with a one-by-one enabling of each
button.