Convert custom Log classes to Kotlin
After reading about SLF4J and trying slf4j-android only to find out
it's deprecated, and reading up on logback-android and finding out
it doesn't work in unit either without Robolectric, and reading up
on logback.qos.ch, and seeing how complicated it is, I settled on
converting my custom Log class to kotlin.  It is so simple but has
no dependencies.