Fix the [net] send logging
The calls to Analytics.logRequest inside XGatewayResponse were too late,
because the get() and post() functions are suspending, and don't return an
HttpResponse until response headers are available.  So to make the logging
timely (and work in the face of timeouts), we do the logging inside the
ktor client plugin.