Squelch fastlane warnings
by using the recommended approach:
    bundle init
    echo 'gem "fastlane"' >> Gemfile
    bundle install

After that I had to add a couple more things to Gemfile
and rerun bundle install to squelch the warnings.
Then run with
    bundle exec fastlane test

or
    tools/fl test

Also it is possible to avoid setting the PATH to include
the JDK and only set JAVA_HOME (as is done in the fl script).
However, that is not enough to make the gradle() action work
in this multi-project workspace.