fastlane: Fix quota error from query_all_version_codes
The Supply client was being reused between calls to
google_play_track_version_codes, returning info for the wrong
project and exhausing the project's request quota.  Merely changing
directories does not clear the cached client and the client was not
present in the lane_context to null it out.  So instead:

* Find all app names with get_all_apps()
* Get package_name, fa_property_id from secret/${app}-info.json
* Pass package_name, json_key params the action

This may be the template for future lanes that iterate.