Add withAsyncBusy to BaseActivity
Use a programmatically generated layout with a styled circular ProgressBar
instead of a ProgressDialog.  Surround account management ops with this
so that the user can't obviously interact with the UI whilst busy.
Actually they can because we do not shortcut the back button, but this
UI is cleaner and I think it can be used everywhere.